move to function.ts
