+++ blog/static/js/functions.ts
@@ -195,3 +195,18 @@ export function extractLinks(xmlDoc: Document): (string | undefined)[] {
   }
   return links;
 }
+
+
+export async function searchLinks(): Promise<string | null> {
+  const key = "eniolakunle_XML";
+  const cacheExists = window.sessionStorage.getItem(key);
+  if (!cacheExists) {
+    console.log("No cache, fetching XML");
+    const xmlText = await fetchXmlData(`${window.location.origin}/sitemap.xml`);
+    const xmlDoc = parseXmlString(xmlText);
+    const links = extractLinks(xmlDoc);
+    window.sessionStorage.setItem(key, JSON.stringify(links));
+  }
+
+  return window.sessionStorage.getItem(key)
+}
