+++ blog/static/js/scripts.js
@@ -1,17 +1,4 @@
-import { fadeTransition, endlessScrolling, fetchXmlData, extractLinks, parseXmlString, } from "./functions.js";
-const key = "eniolakunle_XML";
-async function searchButton() {
-    console.log(window.location);
-    const cacheExists = window.sessionStorage.getItem(key);
-    if (!cacheExists) {
-        console.log("No cache, fetching XML");
-        const xmlText = await fetchXmlData(`${window.location.origin}/sitemap.xml`);
-        const xmlDoc = parseXmlString(xmlText);
-        const links = extractLinks(xmlDoc);
-        window.sessionStorage.setItem(key, JSON.stringify(links));
-    }
-}
-searchButton();
+import { fadeTransition, endlessScrolling, searchLinks } from "./functions.js";
 function toggleMenu() {
     var menu = document.getElementById("menu");
     menu.style.display =
