+++ blog/static/js/scripts.ts
@@ -1,26 +1,10 @@
 import {
   fadeTransition,
   endlessScrolling,
-  fetchXmlData,
-  extractLinks,
-  parseXmlString,
+  searchLinks
 } from "./functions.js";
 
-const key = "eniolakunle_XML";
 
-async function searchButton(): Promise<void> {
-  console.log(window.location);
-  const cacheExists = window.sessionStorage.getItem(key);
-  if (!cacheExists) {
-    console.log("No cache, fetching XML");
-    const xmlText = await fetchXmlData(`${window.location.origin}/sitemap.xml`);
-    const xmlDoc = parseXmlString(xmlText);
-    const links = extractLinks(xmlDoc);
-    window.sessionStorage.setItem(key, JSON.stringify(links));
-  }
-}
-
-searchButton();
 
 function toggleMenu(): void {
   var menu = document.getElementById("menu");
